small cleanup and add bench-quick to Makefile
