+++ crates/benchmarks/Cargo.toml
@@ -2,8 +2,9 @@
 name = "revive-benchmarks"
 version = "0.1.0"
 edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+authors = [
+    "Cyrill Leutwiler <cyrill@parity.io>",
+]
 
 [features]
 default = ["bench-pvm-interpreter"]
