--- crates/benchmarks/EXECUTE.md
@@ -1,18 +1,0 @@
-# Benchmarks
-
-## Table of Contents
-
-- [Benchmark Results](#benchmark-results)
-    - [Baseline](#baseline)
-
-## Benchmark Results
-
-### Baseline
-
-|         | `EVM`                     | `PVMInterpreter`                  |
-|:--------|:--------------------------|:--------------------------------- |
-| **`0`** | `533.12 ns` (✅ **1.00x**) | `471.42 ns` (✅ **1.13x faster**)  |
-
----
-Made with [criterion-table](https://github.com/nu11ptr/criterion-table)
-
