--- crates/benchmarks/src/runtimes/mod.rs
@@ -1,1 +1,0 @@
-pub mod polkavm;
